file rename from snake case to kebab case
